upgrade react and testing again, coz react 17 uses ReactDOM.render() instead of ReactDOM.createRoot() in index.js
The error message indicates that react-dom/client is missing. This module is only available in react-dom version 18 or higher. Since the project is using react@17, this mismatch is causing the build to fail.

Here’s how to fix this:

Option 1: Upgrade to React 18
Since react-dom/client is part of React 18, upgrading react and react-dom to version 18 would solve the issue.